#5606 fixed isOverUnderPayment checkbox
#5606 Payment window calculation of skonto yes/ no when adjusting date
