+++ de.metas.business/src/main/java/de/metas/payment/api/impl/AbstractPaymentDAO.java
@@ -25,6 +25,9 @@ import static org.adempiere.model.InterfaceWrapperHelper.load;
  */
 
 import java.math.BigDecimal;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.sql.Timestamp;
 import java.util.List;
 import java.util.Properties;
@@ -37,13 +40,17 @@ import org.adempiere.ad.dao.IQueryFilter;
 import org.adempiere.ad.dao.IQueryOrderBy;
 import org.adempiere.ad.dao.impl.CompareQueryFilter.Operator;
 import org.adempiere.ad.trx.api.ITrx;
+import org.adempiere.exceptions.DBException;
 import org.adempiere.model.InterfaceWrapperHelper;
 import org.compiere.model.IQuery;
 import org.compiere.model.I_C_AllocationLine;
+import org.compiere.model.I_C_DocType;
 import org.compiere.model.I_C_Invoice;
 import org.compiere.model.I_C_PaySelection;
 import org.compiere.model.I_C_Payment;
 import org.compiere.model.I_Fact_Acct;
+import org.compiere.model.X_C_DocType;
+import org.compiere.util.DB;
 
 import de.metas.adempiere.model.I_C_PaySelectionLine;
 import de.metas.allocation.api.IAllocationDAO;
@@ -180,5 +187,69 @@ public abstract class AbstractPaymentDAO implements IPaymentDAO
 				.create()
 				.listIds(PaymentId::ofRepoId)
 				.stream();
+	}
+	@Override
+	public void updateDiscountAndPayment(I_C_Payment payment,int c_Invoice_ID, I_C_DocType c_DocType)
+	{
+		String sql = "SELECT C_BPartner_ID,C_Currency_ID," // 1..2
+				+ " invoiceOpen(C_Invoice_ID, ?)," // 3 #1
+				+ " invoiceDiscount(C_Invoice_ID,?,?), IsSOTrx " // 4..5 #2/3
+				+ "FROM C_Invoice WHERE C_Invoice_ID=?"; // #4
+		PreparedStatement pstmt = null;
+		ResultSet rs = null;
+		try
+		{
+			pstmt = DB.prepareStatement(sql, null);
+			pstmt.setInt(1, c_Invoice_ID);
+			pstmt.setTimestamp(2, payment.getDateTrx());
+			pstmt.setInt(3, c_Invoice_ID);
+			pstmt.setInt(4, c_Invoice_ID);
+			rs = pstmt.executeQuery();
+			if (rs.next())
+			{
+				final int bpartnerId = rs.getInt(1);
+				payment.setC_BPartner_ID(bpartnerId);
+
+				// Set Invoice Currency
+				final int C_Currency_ID = rs.getInt(2);
+				payment.setC_Currency_ID(C_Currency_ID);
+
+				//
+				BigDecimal InvoiceOpen = rs.getBigDecimal(3); // Set Invoice
+				// OPen Amount
+				if (InvoiceOpen == null)
+				{
+					InvoiceOpen = BigDecimal.ZERO;
+				}
+				BigDecimal DiscountAmt = rs.getBigDecimal(4); // Set Discount
+				// Amt
+				if (DiscountAmt == null)
+				{
+					DiscountAmt = BigDecimal.ZERO;
+				}
+
+				BigDecimal payAmt = InvoiceOpen.subtract(DiscountAmt);
+				if (X_C_DocType.DOCBASETYPE_APCreditMemo.equals(c_DocType.getDocBaseType())
+						|| X_C_DocType.DOCBASETYPE_ARCreditMemo.equals(c_DocType.getDocBaseType()))
+				{
+					if (payAmt.signum() < 0)
+					{
+						payAmt = payAmt.abs();
+					}
+				}
+
+				payment.setPayAmt(payAmt);
+				payment.setDiscountAmt(DiscountAmt);
+			}
+		}
+		catch (SQLException e)
+		{
+			throw new DBException(e, sql);
+		}
+		finally
+		{
+			DB.close(rs, pstmt);
+		}
+		
 	}
 }
